requirements.txt
torch
torchvision
hydra-core
tensorboard
numpy
pandas
omegaconf
Pillow
scikit-learn
captum
matplotlib
scikit-image>=0.19.0